requirements.txt
asgiref==3.8.1
boto3==1.38.15
botocore==1.38.15
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
polyline==1.4.0
geopy==2.4.1
requests==2.32.3